feat(observability): capturer la page d'erreur auth côté serveur
La capture Sentry de /auth/error était uniquement client (useEffect) :
les réseaux d'entreprise/administration qui bloquent sentry.io ne
remontaient rien, même quand l'utilisateur voyait la page « lien
expiré » (incident @interieur.gouv.fr, 8-10 juin).

La page devient un server component qui capture chaque visite avec
user-agent et referer (tag surface: server), l'UI est déplacée dans
le composant client AuthErrorView (webview detection inchangée). La
capture client est supprimée pour éviter le double comptage.
