+++ src/app/[locale]/auth/error/page.tsx
@@ -1,94 +1,35 @@
-"use client";
-
-import { useEffect, useRef, useState } from "react";
-import { useSearchParams } from "next/navigation";
-import { useTranslations } from "next-intl";
 import * as Sentry from "@sentry/nextjs";
-import { Button } from "@/components/ui/button";
-import { Link } from "@/i18n/navigation";
-import { ExternalLink } from "lucide-react";
-import { isInAppBrowser } from "@/lib/detect-webview";
-import { classifyAuthError, AUTH_ERROR_VERIFICATION } from "@/lib/auth/error-kinds";
+import { headers } from "next/headers";
+import { classifyAuthError } from "@/lib/auth/error-kinds";
+import { AuthErrorView } from "./auth-error-view";
 
-function AuthHint({ explanation, action }: { explanation: string; action: string }) {
-  return (
-    <div className="rounded-lg border border-primary/30 bg-primary/5 px-4 py-3 text-sm text-muted-foreground text-left space-y-2">
-      <p>{explanation}</p>
-      <p className="font-medium text-foreground">{action}</p>
-    </div>
-  );
-}
+type Props = {
+  searchParams: Promise<{ error?: string }>;
+};
 
-export default function AuthErrorPage() {
-  const searchParams = useSearchParams();
-  const error = searchParams.get("error");
-  const t = useTranslations("Auth");
-  const [webview, setWebview] = useState(false);
-  const capturedRef = useRef<string | null>(null);
+// Capture côté serveur : les réseaux d'entreprise/administration qui bloquent
+// sentry.io (incident @interieur.gouv.fr) ne remontent jamais une capture
+// client. Le rendu serveur, lui, voit chaque visite de la page avec son
+// user-agent — ce qui permet de distinguer humains et scanners email.
+export default async function AuthErrorPage({ searchParams }: Props) {
+  const { error } = await searchParams;
 
-  useEffect(() => {
-    setWebview(isInAppBrowser());
-  }, []);
+  if (error) {
+    const headersList = await headers();
+    Sentry.captureMessage(`auth-error-page: ${error}`, {
+      level: "warning",
+      tags: {
+        context: "auth",
+        error_code: error,
+        auth_error_kind: classifyAuthError(error),
+        surface: "server",
+      },
+      extra: {
+        user_agent: headersList.get("user-agent") ?? "unknown",
+        referer: headersList.get("referer") ?? "unknown",
+      },
+    });
+  }
 
-  useEffect(() => {
-    if (error && capturedRef.current !== error) {
-      capturedRef.current = error;
-      Sentry.captureMessage(`auth-error-page: ${error}`, {
-        level: "warning",
-        tags: {
-          context: "auth",
-          error_code: error,
-          auth_error_kind: classifyAuthError(error),
-        },
-      });
-    }
-  }, [error]);
-
-  const errorMessage = error
-    ? t(`errors.${error}`, { defaultValue: t("errors.Default") })
-    : t("errors.Default");
-
-  const isVerification = error === AUTH_ERROR_VERIFICATION;
-  const ctaLabel = isVerification ? t("error.requestNewLink") : t("error.backToSignIn");
-
-  return (
-    <div className="flex min-h-screen items-center justify-center px-4">
-      <div className="w-full max-w-sm space-y-4 text-center">
-        <h1 className="text-2xl font-bold tracking-tight">{t("error.title")}</h1>
-        <p className="text-muted-foreground text-sm">{errorMessage}</p>
-
-        {isVerification && (
-          <AuthHint
-            explanation={t("error.verificationExplanation")}
-            action={t("error.verificationAction")}
-          />
-        )}
-
-        {webview && (
-          <AuthHint
-            explanation={t("error.webviewExplanation")}
-            action={t("error.webviewAction")}
-          />
-        )}
-
-        <div className="flex flex-col gap-2">
-          <Button asChild>
-            <Link href="/auth/sign-in">{ctaLabel}</Link>
-          </Button>
-          {webview && (
-            <Button variant="outline" asChild>
-              <a
-                href={typeof window !== "undefined" ? window.location.origin : "https://the-playground.fr"}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <ExternalLink className="size-4" />
-                {t("error.openInBrowser")}
-              </a>
-            </Button>
-          )}
-        </div>
-      </div>
-    </div>
-  );
+  return <AuthErrorView error={error ?? null} />;
 }
